bot/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==1.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
cytoolz==0.11.2
dnspython==2.2.1
docopt==0.6.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
ffmpeg==1.4
Flask==2.1.1
fonttools==4.33.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.5.1
kiwisolver==1.4.2
lru-dict==1.1.7
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
midi2audio==0.1.1
more-itertools==8.12.0
motor==2.5.1
multiaddr==0.0.9
multidict==6.0.2
music21==7.3.1
netaddr==0.8.0
numpy==1.22.3
packaging==21.3
parsimonious==0.8.1
Pillow==9.1.0
pipreqs==0.4.11
protobuf==3.20.1
psutil==5.9.0
git+https://github.com/Pycord-Development/pycord.git
pycryptodome==3.14.1
pymongo==3.12.3
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
soupsieve==2.3.2.post1
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.1
webcolors==1.11.1
websockets==9.1
Werkzeug==2.1.1
yarg==0.1.9
yarl==1.7.2
